Send session-request before receiving bind-response.
This modification sends the bind and session requests in succession without
waiting for response to the former.

This fixes session establishment to xmpp.slack.com.  The problem is that we
receive a version request before the session has been established.  If we
register the version-extension after establishing the session, as we would
normally do, then we fail with a StreamError upon receiving this request.
If we register the version-extension before establishing the session, then
we also respond to it before establishing the session, which causes slack to
send repeated version-requests until the session establishment times out.